fix for empty output

+++ CI/clang_tidy/run_clang_tidy.sh
@@ -11,12 +11,13 @@ mkdir -p $output_dir
 output_dir=$(realpath $output_dir)
 
 pushd $build_dir
-echo "why?"
-NINJA_STATUS="[ninja][%f/%t] " ninja $@ | tee $output_dir/ninja.log
-echo "what?"
+NINJA_STATUS="[ninja] [%f/%t] " ninja $@ | tee $output_dir/ninja.log
 popd
-echo "hi!"
-cat $output_dir/ninja.log | grep -v '\[ninja\]' > $output_dir/clang-tidy.log
 
-echo "hello?"
-exit 0
+# grep "fails" for empty input
+if [ ! -s $output_dir/ninja.log ]; then
+  touch $output_dir/clang-tidy.log
+  exit 0
+fi
+
+cat $output_dir/ninja.log | grep -v '\[ninja\]' > $output_dir/clang-tidy.log
